--story=876554553 cos分片下载时，优化降级逻辑，更快速响应请求

+++ src/backend/common/common-storage/storage-service/src/main/kotlin/com/tencent/bkrepo/common/storage/innercos/request/FileCleanupChunkedFutureListener.kt
@@ -8,7 +8,10 @@ import org.apache.commons.logging.LogFactory
 
 class FileCleanupChunkedFutureListener : ChunkedFutureListener<File> {
     override fun done(future: Future<File>?, getInputStreamTime: Long) {
-        val file = future?.get() ?: return
+        if (future == null || future.isCancelled) {
+            return
+        }
+        val file = future.get()
         try {
             Files.deleteIfExists(file.toPath())
             logger.info("Delete cos downloading temp file[$file] success.")
